refactor: Separate views for Customer & Owner
